refactor: enum값 converter 기능으로 리팩터링
