+++ backend/src/main/java/com/now/naaga/common/config/WebConfig.java
@@ -6,8 +6,12 @@ import com.now.naaga.auth.presentation.interceptor.AuthInterceptor;
 import com.now.naaga.auth.presentation.interceptor.ManagerAuthInterceptor;
 import com.now.naaga.common.presentation.interceptor.RequestMatcherInterceptor;
 import java.util.List;
+
+import com.now.naaga.letter.domain.LetterLogTypeConverter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.format.FormatterRegistrar;
+import org.springframework.format.FormatterRegistry;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerInterceptor;
@@ -37,6 +41,11 @@ public class WebConfig implements WebMvcConfigurer {
         this.memberAuthArgumentResolver = memberAuthArgumentResolver;
         this.authInterceptor = authInterceptor;
         this.managerAuthInterceptor = managerAuthInterceptor;
+    }
+
+    @Override
+    public void addFormatters(final FormatterRegistry registry){
+        registry.addConverter(new LetterLogTypeConverter());
     }
 
     @Override
