+++ backend/src/main/java/com/now/naaga/letter/presentation/LetterController.java
@@ -1,7 +1,6 @@
 package com.now.naaga.letter.presentation;
 
 import com.now.naaga.auth.presentation.annotation.Auth;
-import com.now.naaga.common.exception.CommonException;
 import com.now.naaga.letter.application.LetterService;
 import com.now.naaga.letter.application.dto.CreateLetterCommand;
 import com.now.naaga.letter.domain.Letter;
@@ -14,11 +13,7 @@ import org.springframework.web.bind.annotation.*;
 import java.net.URI;
 import java.util.List;
 
-import static com.now.naaga.common.exception.CommonExceptionType.INVALID_REQUEST_PARAMETERS;
-import static com.now.naaga.letter.presentation.LogType.READ;
-import static com.now.naaga.letter.presentation.LogType.WRITE;
-
-@RequestMapping("/letters")
+@RequestMapping
 @RestController
 public class LetterController {
 
@@ -65,15 +60,12 @@ public class LetterController {
     @GetMapping("/letterlogs")
     public ResponseEntity<List<LetterResponse>> findLetterInGame(@Auth final PlayerRequest playerRequest,
                                                                  @RequestParam final Long gameId,
-                                                                 @RequestParam final String logType) {
-        if (!(READ.name().equalsIgnoreCase(logType) || WRITE.name().equalsIgnoreCase(logType))) {
-            throw new CommonException(INVALID_REQUEST_PARAMETERS);
-        }
+                                                                 @RequestParam final LetterLogType logType) {
         final FindLetterLogByGameCommand findLetterLogByGameCommand = FindLetterLogByGameCommand.of(playerRequest, gameId, logType);
         final List<Letter> letters = letterService.findLetterLogInGame(findLetterLogByGameCommand);
 
         final List<LetterResponse> writeLetterResponses = letters.stream()
-                .map(letter -> LetterResponse.from(letter))
+                .map(LetterResponse::from)
                 .toList();
         return ResponseEntity
                 .status(HttpStatus.OK)
