--- backend/src/main/java/com/now/naaga/letter/presentation/LogType.java
@@ -1,8 +1,0 @@
-package com.now.naaga.letter.presentation;
-
-public enum LogType {
-
-    READ,
-    WRITE,
-    ;
-}
