+++ backend/src/main/java/com/now/naaga/letter/presentation/dto/FindLetterLogByGameCommand.java
@@ -1,16 +1,16 @@
 package com.now.naaga.letter.presentation.dto;
 
-import com.now.naaga.letter.presentation.LogType;
+import com.now.naaga.letter.presentation.LetterLogType;
 import com.now.naaga.player.presentation.dto.PlayerRequest;
 
 public record FindLetterLogByGameCommand(Long playerId,
                                          Long gameId,
-                                         LogType logType) {
+                                         LetterLogType letterLogType) {
 
     public static FindLetterLogByGameCommand of(final PlayerRequest playerRequest,
                                                 final Long gameId,
-                                                final String logType) {
-        return new FindLetterLogByGameCommand(playerRequest.playerId(), gameId, LogType.valueOf(logType));
+                                                final LetterLogType logType) {
+        return new FindLetterLogByGameCommand(playerRequest.playerId(), gameId, logType);
     }
 }
