+++ backend/src/test/java/com/now/naaga/letter/presentation/LetterControllerTest.java
@@ -4,16 +4,15 @@ import com.now.naaga.auth.domain.AuthToken;
 import com.now.naaga.auth.infrastructure.AuthType;
 import com.now.naaga.auth.infrastructure.jwt.AuthTokenGenerator;
 import com.now.naaga.common.CommonControllerTest;
-import com.now.naaga.common.builder.GameBuilder;
-import com.now.naaga.common.builder.LetterBuilder;
-import com.now.naaga.common.builder.PlaceBuilder;
-import com.now.naaga.common.builder.PlayerBuilder;
+import com.now.naaga.common.builder.*;
 import com.now.naaga.common.exception.ExceptionResponse;
 import com.now.naaga.game.domain.Game;
 import com.now.naaga.game.presentation.dto.CoordinateResponse;
 import com.now.naaga.game.repository.GameRepository;
 import com.now.naaga.letter.application.LetterService;
 import com.now.naaga.letter.domain.Letter;
+import com.now.naaga.letter.domain.letterlog.ReadLetterLog;
+import com.now.naaga.letter.domain.letterlog.WriteLetterLog;
 import com.now.naaga.letter.presentation.dto.LetterRequest;
 import com.now.naaga.letter.presentation.dto.LetterResponse;
 import com.now.naaga.letter.presentation.dto.NearByLetterResponse;
@@ -35,11 +34,11 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
+import static com.now.naaga.common.exception.CommonExceptionType.INVALID_REQUEST_PARAMETERS;
 import static com.now.naaga.common.fixture.PositionFixture.*;
 import static com.now.naaga.game.domain.GameStatus.DONE;
 import static com.now.naaga.game.exception.GameExceptionType.NOT_EXIST_IN_PROGRESS;
@@ -49,102 +48,108 @@ import static org.assertj.core.api.SoftAssertions.assertSoftly;
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class LetterControllerTest extends CommonControllerTest {
-    
+
     @Autowired
     private AuthTokenGenerator authTokenGenerator;
-    
+
     @Autowired
     private GameRepository gameRepository;
-    
+
     @Autowired
     private LetterService letterService;
-    
+
     @Autowired
     private LetterRepository letterRepository;
-    
+
     @Autowired
     private GameBuilder gameBuilder;
-    
+
     @Autowired
     private LetterBuilder letterBuilder;
-    
+
+    @Autowired
+    private ReadLetterLogBuilder readLetterLogBuilder;
+
+    @Autowired
+    private WriteLetterLogBuilder writeLetterLogBuilder;
+
     @Autowired
     private PlaceBuilder placeBuilder;
-    
+
     @Autowired
     private PlayerBuilder playerBuilder;
-    
+
     @BeforeEach
     protected void setUp() {
         super.setUp();
     }
-    
+
     @Test
     void 주변_쪽지를_모두_조회한다() {
         // given
         final Place destination = placeBuilder.init()
-                                              .build();
+                .build();
         final Player player = playerBuilder.init()
-                                           .build();
+                .build();
         final Letter letter1 = letterBuilder.init()
-                                            .build();
+                .build();
         final Letter letter2 = letterBuilder.init()
-                                            .position(잠실역_교보문고_110미터_앞_좌표)
-                                            .build();
-        
+                .position(잠실역_교보문고_110미터_앞_좌표)
+                .build();
+
         final AuthToken generate = authTokenGenerator.generate(destination.getRegisteredPlayer().getMember(), 1L, AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         // when
         final ExtractableResponse<Response> extract = RestAssured.given().log().all()
-                                                                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                                                                 .header("Authorization", "Bearer " + accessToken)
-                                                                 .param("latitude", 잠실역_교보문고_좌표.getLatitude())
-                                                                 .param("longitude", 잠실역_교보문고_좌표.getLongitude())
-                                                                 .when()
-                                                                 .get("/letters/nearby")
-                                                                 .then().log().all()
-                                                                 .statusCode(HttpStatus.OK.value())
-                                                                 .extract();
-        
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .header("Authorization", "Bearer " + accessToken)
+                .param("latitude", 잠실역_교보문고_좌표.getLatitude())
+                .param("longitude", 잠실역_교보문고_좌표.getLongitude())
+                .when()
+                .get("/letters/nearby")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
         // then
         final List<NearByLetterResponse> actual = extract.as(new TypeRef<>() {
         });
         final int statusCode = extract.statusCode();
         final List<NearByLetterResponse> expected = List.of(NearByLetterResponse.from(letter1));
-        
+
         assertSoftly(softly -> {
             softly.assertThat(statusCode).isEqualTo(HttpStatus.OK.value());
             softly.assertThat(actual)
-                  .hasSize(1)
-                  .usingRecursiveComparison()
-                  .isEqualTo(expected);
+                    .hasSize(1)
+                    .usingRecursiveComparison()
+                    .isEqualTo(expected);
         });
     }
-    
+
     @Test
     void 쪽지_식별자로_쪽지를_조회한다() {
         // given & when
         final Player player = playerBuilder.init()
-                                           .build();
-        
+                .build();
+
         final Place destination = placeBuilder.init()
-                                              .position(잠실_루터회관_정문_좌표)
-                                              .build();
-        
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
         final Game game = gameBuilder.init()
-                                     .place(destination)
-                                     .player(player)
-                                     .startPosition(잠실역_교보문고_좌표)
-                                     .build();
-        
+                .place(destination)
+                .player(player)
+                .startPosition(잠실역_교보문고_좌표)
+                .build();
+
         final Letter letter = letterBuilder.init()
-                                           .position(잠실_루터회관_정문_좌표)
-                                           .build();
-        
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
         final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         final ExtractableResponse<Response> extract = RestAssured
                 .given().log().all()
                 .header("Authorization", "Bearer " + accessToken)
@@ -152,39 +157,39 @@ class LetterControllerTest extends CommonControllerTest {
                 .get("/letters/{letterId}", letter.getId())
                 .then().log().all()
                 .extract();
-        
+
         // then
         final int statusCode = extract.statusCode();
         final LetterResponse actual = extract.as(LetterResponse.class);
         final LetterResponse expected = LetterResponse.from(letter);
-        
+
         assertSoftly(softAssertions -> {
             softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.OK.value());
             softAssertions.assertThat(actual)
-                          .usingRecursiveComparison()
-                          .ignoringExpectedNullFields()
-                          .ignoringFields("registerDate")
-                          .isEqualTo(expected);
+                    .usingRecursiveComparison()
+                    .ignoringExpectedNullFields()
+                    .ignoringFields("registerDate")
+                    .isEqualTo(expected);
         });
     }
-    
+
     @Test
     void 쪽지_식별자로_쪽지를_조회시_잔행중_게임이_없으면_예외가_발생한다() {
         // given & when
         final Player player = playerBuilder.init()
-                                           .build();
-        
+                .build();
+
         final Place destination = placeBuilder.init()
-                                              .position(잠실_루터회관_정문_좌표)
-                                              .build();
-        
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
         final Letter letter = letterBuilder.init()
-                                           .position(잠실_루터회관_정문_좌표)
-                                           .build();
-        
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
         final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         final ExtractableResponse<Response> extract = RestAssured
                 .given().log().all()
                 .header("Authorization", "Bearer " + accessToken)
@@ -192,40 +197,40 @@ class LetterControllerTest extends CommonControllerTest {
                 .get("/letters/{letterId}", letter.getId())
                 .then().log().all()
                 .extract();
-        
+
         // then
         final int statusCode = extract.statusCode();
         final ExceptionResponse actual = extract.as(ExceptionResponse.class);
-        
+
         final ExceptionResponse expected = new ExceptionResponse(NOT_EXIST_IN_PROGRESS.errorCode(), NOT_EXIST_IN_PROGRESS.errorMessage());
-        
+
         assertSoftly(softAssertions -> {
             softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.NOT_FOUND.value());
             softAssertions.assertThat(actual)
-                          .usingRecursiveComparison()
-                          .ignoringExpectedNullFields()
-                          .ignoringFieldsOfTypes(LocalDateTime.class)
-                          .isEqualTo(expected);
+                    .usingRecursiveComparison()
+                    .ignoringExpectedNullFields()
+                    .ignoringFieldsOfTypes(LocalDateTime.class)
+                    .isEqualTo(expected);
         });
     }
-    
+
     @Test
     void 쪽지_등록_요청시_쪽지를_등록한_플레이어가_존재하고_그_플레이어가_진행중인_게임이_있으면_정상적으로_쪽지를_생성한다() {
         //given
         final Player player = playerBuilder.init()
-                                           .build();
+                .build();
         final Game game = gameBuilder.init()
-                                     .player(player)
-                                     .build();
-        
+                .player(player)
+                .build();
+
         final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         String message = "날씨가 선선하네요";
         final LetterRequest letterRequest = new LetterRequest(message,
                 잠실_루터회관_정문_좌표.getLatitude().doubleValue(),
                 잠실_루터회관_정문_좌표.getLongitude().doubleValue());
-        
+
         //when
         final ExtractableResponse<Response> extract = RestAssured
                 .given().log().all()
@@ -236,7 +241,7 @@ class LetterControllerTest extends CommonControllerTest {
                 .post("/letters")
                 .then().log().all()
                 .extract();
-        
+
         //then
         final int statusCode = extract.statusCode();
         final String location = extract.header("Location");
@@ -249,18 +254,18 @@ class LetterControllerTest extends CommonControllerTest {
                 message,
                 null
         );
-        
+
         assertSoftly(softAssertions -> {
                     softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.CREATED.value());
                     softAssertions.assertThat(location).isNotEqualTo("/letters" + letterId);
                     softAssertions.assertThat(actual)
-                                  .usingRecursiveComparison()
-                                  .ignoringExpectedNullFields()
-                                  .isEqualTo(expected);
+                            .usingRecursiveComparison()
+                            .ignoringExpectedNullFields()
+                            .isEqualTo(expected);
                 }
         );
     }
-    
+
     @Test
     void 쪽지_등록_요청시_쪽지를_등록한_플레이어가_존재하지_않으면_예외가_발생한다() {
         //given
@@ -268,12 +273,12 @@ class LetterControllerTest extends CommonControllerTest {
                 1L,
                 AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         String message = "날씨가 선선하네요";
         final LetterRequest letterRequest = new LetterRequest(message,
                 잠실_루터회관_정문_좌표.getLatitude().doubleValue(),
                 잠실_루터회관_정문_좌표.getLongitude().doubleValue());
-        
+
         //when
         RestAssured.defaultParser = Parser.JSON;
         final ExtractableResponse<Response> extract = RestAssured
@@ -285,7 +290,7 @@ class LetterControllerTest extends CommonControllerTest {
                 .post("/letters")
                 .then().log().all()
                 .extract();
-        
+
         //then
         final int statusCode = extract.statusCode();
         final ExceptionResponse actual = extract.as(ExceptionResponse.class);
@@ -293,35 +298,35 @@ class LetterControllerTest extends CommonControllerTest {
                 PLAYER_NOT_FOUND.errorCode(),
                 PLAYER_NOT_FOUND.errorMessage()
         );
-        
+
         assertSoftly(softAssertions -> {
                     softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.NOT_FOUND.value());
                     softAssertions.assertThat(actual)
-                                  .usingRecursiveComparison()
-                                  .isEqualTo(expected);
+                            .usingRecursiveComparison()
+                            .isEqualTo(expected);
                 }
         );
     }
-    
+
     @Test
     void 쪽지_등록_요청시_쪽지를_등록한_진행중인_게임이_존재하지_않으면_예외가_발생한다() {
         //given
         final Player player = playerBuilder.init()
-                                           .build();
+                .build();
         final Game finishedGame = gameBuilder.init()
-                                             .player(player)
-                                             .gameStatus(DONE)
-                                             .build();
+                .player(player)
+                .gameStatus(DONE)
+                .build();
         final AuthToken generate = authTokenGenerator.generate(player.getMember(),
                 1L,
                 AuthType.KAKAO);
         final String accessToken = generate.getAccessToken();
-        
+
         String message = "날씨가 선선하네요";
         final LetterRequest letterRequest = new LetterRequest(message,
                 잠실_루터회관_정문_좌표.getLatitude().doubleValue(),
                 잠실_루터회관_정문_좌표.getLongitude().doubleValue());
-        
+
         //when
         RestAssured.defaultParser = Parser.JSON;
         final ExtractableResponse<Response> extract = RestAssured
@@ -333,7 +338,7 @@ class LetterControllerTest extends CommonControllerTest {
                 .post("/letters")
                 .then().log().all()
                 .extract();
-        
+
         //then
         final int statusCode = extract.statusCode();
         final ExceptionResponse actual = extract.as(ExceptionResponse.class);
@@ -341,12 +346,204 @@ class LetterControllerTest extends CommonControllerTest {
                 NOT_EXIST_IN_PROGRESS.errorCode(),
                 NOT_EXIST_IN_PROGRESS.errorMessage()
         );
-        
+
         assertSoftly(softAssertions -> {
                     softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.NOT_FOUND.value());
                     softAssertions.assertThat(actual)
-                                  .usingRecursiveComparison()
-                                  .isEqualTo(expected);
+                            .usingRecursiveComparison()
+                            .isEqualTo(expected);
+                }
+        );
+    }
+
+    @Test
+    void 게임아이디와_회원으로_읽은쪽지로그를_모두_조회한다() {
+        // given & when
+        final Player player = playerBuilder.init()
+                .build();
+
+        final Place destination = placeBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Game game1 = gameBuilder.init()
+                .place(destination)
+                .player(player)
+                .startPosition(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Game game2 = gameBuilder.init()
+                .place(destination)
+                .player(player)
+                .startPosition(역삼역_좌표)
+                .build();
+
+        final Letter letter = letterBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final ReadLetterLog readLetterLog1 = readLetterLogBuilder.init()
+                .game(game1)
+                .letter(letter)
+                .build();
+
+        final ReadLetterLog readLetterLog2 = readLetterLogBuilder.init()
+                .game(game2)
+                .letter(letter)
+                .build();
+
+        final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
+        final String accessToken = generate.getAccessToken();
+
+        final ExtractableResponse<Response> extract = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
+                .param("gameId", game1.getId())
+                .param("logType", "READ")
+                .when()
+                .get("/letterlogs")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        // then
+        final List<LetterResponse> actual = extract.as(new TypeRef<>() {
+        });
+        final int statusCode = extract.statusCode();
+        final List<LetterResponse> expected = List.of(LetterResponse.from(letter));
+
+        assertSoftly(softly -> {
+            softly.assertThat(statusCode).isEqualTo(HttpStatus.OK.value());
+            softly.assertThat(actual)
+                    .hasSize(1)
+                    .usingRecursiveComparison()
+                    .ignoringFields("registerDate")
+                    .isEqualTo(expected);
+        });
+    }
+
+    @Test
+    void
+    게임아이디와_회원으로_작성한_쪽지로그를_모두_조회한다() {
+        // given & when
+        final Player player = playerBuilder.init()
+                .build();
+
+        final Place destination = placeBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Game game = gameBuilder.init()
+                .place(destination)
+                .player(player)
+                .startPosition(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Letter letter = letterBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final WriteLetterLog writeLetterLog1 = writeLetterLogBuilder.init()
+                .game(game)
+                .letter(letter)
+                .build();
+
+        final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
+        final String accessToken = generate.getAccessToken();
+
+        final ExtractableResponse<Response> extract = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
+                .param("gameId", game.getId())
+                .param("logType", "WRITE")
+                .when()
+                .get("/letterlogs")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract();
+
+        // then
+        final List<LetterResponse> actual = extract.as(new TypeRef<>() {
+        });
+        final int statusCode = extract.statusCode();
+        final List<LetterResponse> expected = List.of(LetterResponse.from(letter));
+
+        assertSoftly(softly -> {
+            softly.assertThat(statusCode).isEqualTo(HttpStatus.OK.value());
+            softly.assertThat(actual)
+                    .hasSize(1)
+                    .usingRecursiveComparison()
+                    .ignoringFields("registerDate")
+                    .isEqualTo(expected);
+        });
+    }
+
+    @Test
+    void 쪽지로그_조회시_잘못된_파라미터는_예외를_던진다() {
+        // given & when
+        final Player player = playerBuilder.init()
+                .build();
+
+        final Place destination = placeBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Game game1 = gameBuilder.init()
+                .place(destination)
+                .player(player)
+                .startPosition(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Game game2 = gameBuilder.init()
+                .place(destination)
+                .player(player)
+                .startPosition(역삼역_좌표)
+                .build();
+
+        final Letter letter = letterBuilder.init()
+                .position(잠실_루터회관_정문_좌표)
+                .build();
+
+        final Letter letter2 = letterBuilder.init()
+                .position(GS25_방이도곡점_좌표)
+                .build();
+
+        final WriteLetterLog writeLetterLog1 = writeLetterLogBuilder.init()
+                .game(game1)
+                .letter(letter)
+                .build();
+
+        final WriteLetterLog writeLetterLog2 = writeLetterLogBuilder.init()
+                .game(game2)
+                .letter(letter2)
+                .build();
+
+        final AuthToken generate = authTokenGenerator.generate(player.getMember(), 1L, AuthType.KAKAO);
+        final String accessToken = generate.getAccessToken();
+
+        final ExtractableResponse<Response> extract = RestAssured
+                .given().log().all()
+                .header("Authorization", "Bearer " + accessToken)
+                .param("gameId", game1.getId())
+                .param("logType", "잘못된enum")
+                .when()
+                .get("/letterlogs")
+                .then().log().all()
+                .extract();
+
+        // then
+        final int statusCode = extract.statusCode();
+        final ExceptionResponse actual = extract.as(ExceptionResponse.class);
+        final ExceptionResponse expected = new ExceptionResponse(INVALID_REQUEST_PARAMETERS.errorCode(), INVALID_REQUEST_PARAMETERS.errorMessage());
+
+        assertSoftly(softAssertions -> {
+                    softAssertions.assertThat(statusCode).isEqualTo(HttpStatus.BAD_REQUEST.value());
+                    softAssertions.assertThat(actual)
+                            .usingRecursiveComparison()
+                            .ignoringExpectedNullFields()
+                            .ignoringFieldsOfTypes(LocalDateTime.class)
+                            .ignoringFields("registerDate")
+                            .isEqualTo(expected);
                 }
         );
     }
