+++ backend/src/test/java/com/now/naaga/letter/presentation/LetterLogTypeTest.java
@@ -1,0 +1,11 @@
+package com.now.naaga.letter.presentation;
+
+import org.junit.jupiter.api.Test;
+
+class LetterLogTypeTest {
+
+    @Test
+    void ghgh() {
+        LetterLogType.from("하하하");
+    }
+}
